requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
blinker==1.9.0
cached-property==2.0.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
commonmark==0.9.1
defusedxml==0.7.1
Deprecated==1.2.15
Flask==2.3.2
Flask-Limiter==2.9.2
frozenlist==1.5.0
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.14.1
Markdown==3.7
MarkupSafe==3.0.2
multidict==6.1.0
mypy-extensions==1.0.0
ollama==0.4.5
openai==0.27.0
ordered-set==4.1.0
packaging==24.2
propcache==0.2.1
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
python-dotenv==1.0.1
pytube==15.0.0
regex==2024.11.6
requests==2.32.3
rich==12.6.0
setuptools==75.6.0
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlalchemy-orm==1.2.10
tiktoken==0.8.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.17.0
yarl==1.18.3
youtube-transcript-api==0.6.3
yt-dlp==2024.12.6
PyJWT
